Class Library Ref and misc Blazor pages
