+++ BlazorWebApp/Pages/AddRT.razor
@@ -1,0 +1,20 @@
+@page "/AddRT"
+@inherits AddRTBase
+@* using BlazorWebApp.Models *@
+
+
+<PageTitle>Add your memory!</PageTitle>
+
+<h1>Please add a memory you would like to remember and share!</h1>
+
+<input @bind="Item.rtUserName" placeholder="please add your user name" />
+<br>
+<br>
+<input @bind="Item.rtDescription" placeholder="describe the fun time" />
+
+
+<br>
+
+
+
+
