+++ BlazorWebApp/Pages/Bobo.razor
@@ -5,12 +5,16 @@
 
 <h1>Hello, from Bobo!</h1>
 
-Welcome to a Blazor component web page.
+<input @bind="UserName" placeholder="please add your user name" />
+<br>
+<br>
+<input @bind="Title" placeholder="describe the fun time" />
 
 <br>
 <br>
 <br>
 
-<Counter incrementCount=5/>
+
+
