+++ BlazorWebApp/Pages/Bobo.razor.cs
@@ -1,0 +1,12 @@
+namespace BlazorWebApp.Pages
+{
+ public partial class Bobo
+ {
+    
+    private string UserName = "";
+    private string Title = "Partial Private";
+
+ }
+
+
+}
