+++ BlazorWebApp/Shared/NavMenu.razor
@@ -24,6 +24,11 @@
                 <span class="oi oi-list-rich" aria-hidden="true"></span> Fetch data
             </NavLink>
         </div>
+        <div class="nav-item px-3">
+            <NavLink class="nav-link" href="AddRT">
+                <span class="oi oi-list-rich" aria-hidden="true"></span> Remember This
+            </NavLink>
+        </div>
     </nav>
 </div>
