+++ rtWebApp/Pages/CallAPI.cshtml
@@ -30,8 +30,8 @@
                     <label for="baseUrl">@Model.apiBase</label>
                     <input type="text" asp-for="apiRoute" placeholder="Cocktail"><br>
                     <br>
-                    <label for="thisrtItem.rtUser">User Name: </label>
-                    <input type="text" asp-for="thisrtItem.rtUser" placeholder="user"><br>
+                    <label for="thisrtItem.rtUserName">User Name: </label>
+                    <input type="text" asp-for="thisrtItem.rtUserName" placeholder="user"><br>
                     <br>
                     <label for="thisrtItem.rtLocation">Location : </label>
                     <input type="text" asp-for="thisrtItem.rtLocation" placeholder="location"><br>
