+++ rtWebApp/Pages/CallAPI.cshtml.cs
@@ -1,7 +1,7 @@
 ﻿using System.Net.Http.Headers;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
-using rtWebApp.Models;
+using SharedModels;
 
 
 
@@ -29,7 +29,7 @@ public class CallAPIModel : PageModel
     public IFormFile Upload { get; set; } = null!;
     
     [BindProperty]
-    public rtItem thisrtItem { get; set; } = new rtItem{ rtId = 1001, rtUser = "Cosmo", rtDescription = "fun time digging hole for bone",rtLocation = "backyard", rtDateTime =  DateTime.UtcNow };
+    public rtItem thisrtItem { get; set; } = new rtItem{ rtId = 1001, rtUserName = "Cosmo", rtDescription = "fun time digging hole for bone",rtLocation = "backyard", rtDateTime =  DateTime.UtcNow };
 
     public void OnGet()
     {
